servers/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-dotenv==1.0.1
python-multipart==0.0.9
langchain==0.1.11
langchain-google-genai==0.0.11
pdfplumber==0.11.0
duckduckgo-search==5.2.1
